removes legacy compatibility turbine peers shuffle code
