+++ gossip/src/lib.rs
@@ -13,7 +13,7 @@ pub mod crds_gossip_pull;
 pub mod crds_gossip_push;
 pub mod crds_shards;
 pub mod crds_value;
-pub mod deprecated;
+mod deprecated;
 pub mod duplicate_shred;
 pub mod epoch_slots;
 pub mod gossip_error;
